Implement the idea of the local user
The local user is marked as such when the code for adding a student is
run. In the future, this will most likely use the OAuth access token to
identify the current user.

This commit will block the input box if it is not the students turn to
write.

See #14

See empirical-org/Compass#269