Catch uncaughtException
Typically, in real world code, this is a _horrible_ thing to do.
But this is for the gulp process. The issue is the file gets deleted as
something is trying to write to it. The race condition is solvable, but
that is a longer fix than this quick dirty thing to do.

Closes #67