Properly pipe vendor css and js
The issue with bower main files is that main files can be anything! Not
just javascript. Well of course this makes sense since bower packages
include all assets.

What we need to do is properly extract all the .css .js .* and put them
into the proper built source and inject.