Add angular-local-storage
See #98

Angular local storage is an abstraction using the $localStorage module,
but it provides cookie storage as a backup. This will be good for
browsers that do not implement the localStorage protocol. Which isn't
many these days, but it is good to be safe.

This method will ultimately fail if the user's browser doesn't support
localstorage and cookies are turned off.